+++ contrib/libs/backward/backward.cpp
@@ -37,6 +37,6 @@
 
 namespace backward {
 
-backward::SignalHandling sh;
+// backward::SignalHandling sh;
 
 } // namespace backward