+++ src/modules/io/Filesystem.cpp
@@ -241,6 +241,7 @@ bool Filesystem::list(const core::String &directory, core::DynamicArray<Filesyst
 }
 
 bool Filesystem::chdir(const core::String &directory) {
+	Log::debug("Change current working dir to %s", directory.c_str());
 	return fs_chdir(directory.c_str());
 }